+++ apps/monitor-web/src/queries/login/login.queries.ts
@@ -2,7 +2,7 @@ import { useNavigate, useLocation } from "react-router-dom";
 import { supabase } from "@/lib/supabase";
 import { useToast } from "@/hooks";
 import useAppMutation from "../base/useAppMutation";
-import { LoginFormValues } from "@/pages/Login/types";
+import { LoginFormValues } from "@/pages/Login/_types";
 import { useQueryClient } from "@tanstack/react-query";
 import { authQueryKeys } from "../queryKeys";
